Revert "Add player state pattern"
This reverts commit bd2261a0ff80c8afeab90083720e27fef78070a4.

+++ Assets/Script/Weapon.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public abstract class Weapon : MonoBehaviour
 {
-    [SerializeField] public float Damage;
+    [SerializeField]public int Damage;
     [SerializeField] protected float FireRate;
     protected bool isReady;
     public abstract void Attack();
